test: try splitting out hmr test into separate test file
