+++ test/utils.ts
@@ -1,9 +1,6 @@
-import { fileURLToPath } from 'node:url'
 import { expect } from 'vitest'
 import type { Page } from 'playwright'
 import { createPage, getBrowser, url, useTestContext } from '@nuxt/test-utils'
-
-export const fixturesDir = fileURLToPath(new URL(process.env.TEST_ENV === 'dev' ? './fixtures-temp' : './fixtures', import.meta.url))
 
 export async function renderPage (path = '/') {
   const ctx = useTestContext()
